meta-lxatac-software: lxatac-core-bundle-base: accept empty migrate.d dir
Prior to this patch the part of the hook script that handles migration of
user-specified files will fail if the `/etc/rauc/migrate.d` directory
exists but does not contain any `*.conf` files.

Make sure that the glob expands to nothing instead of the string
`/etc/rauc/migrate.d/*.conf` in this case by setting the `nullglob`
option. This also removes the need for the ineffective check for the
existance of the `/etc/rauc/migrate.d` directory.

This also affects the other for loop with a glob, which searches for
available certificate files. The behaviour of this loop also matches the
expected behaviour better with the `nullglob` option enabled.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>